Work on stopping all LinkSet and Link timers
It has been a while. Most of the machines are specified in Q.704 and Q.707 but they don't seem to be properly integrated. When the link test has failed and the link is in-active stop the linktest and keep the link down. It would be the task of the LinkSet machines to recover the link.

But actually we will have to see.