Avoid overloading the queue
In case we receive more SCCP messages than we can handle we need to throttle. This could be a proper congestion handling in the future but for now just drop them. This way we make sure we don't respond to very old items.  Print a prominent log message that shows the data dropped and the amount of pending blocks in the system.